git: Allow signing of commits
We could add more ways of specifying a key, such as picking one from the
adopted identity, but for now with this change we allow users who know what
they are doing to sign their bug changes.

By simply passing "-S" to `git commit-tree` when "commit.gpgsign" is "true", we
use git's mechanism for key choosing which is either by the configured user
name and email OR by whatever is specified in "user.signingkey".